+++ examples/with-hasura-thirdpartyemailpassword/api-server.js
@@ -1,0 +1,105 @@
+const express = require("express");
+const cors = require("cors");
+const helmet = require("helmet");
+require("dotenv").config();
+let supertokens = require("supertokens-node");
+let Session = require("supertokens-node/recipe/session");
+let { middleware, errorHandler } = require("supertokens-node/framework/express");
+let ThirdPartyEmailPassword = require("supertokens-node/recipe/thirdpartyemailpassword");
+
+const apiPort = process.env.REACT_APP_API_PORT || 3001;
+const apiDomain = process.env.REACT_APP_API_URL || `http://localhost:${apiPort}`;
+const websitePort = process.env.REACT_APP_WEBSITE_PORT || 3000;
+const websiteDomain = process.env.REACT_APP_WEBSITE_URL || `http://localhost:${websitePort}`;
+
+supertokens.init({
+    framework: "express",
+    supertokens: {
+        connectionURI: "https://try.supertokens.com",
+    },
+    appInfo: {
+        appName: "SuperTokens Hasura Demo App",
+        apiDomain,
+        websiteDomain,
+    },
+    recipeList: [
+        ThirdPartyEmailPassword.init({
+            providers: [
+                // We have provided you with development keys which you can use for testing.
+                // IMPORTANT: Please replace them with your own OAuth keys for production use.
+                ThirdPartyEmailPassword.Google({
+                    clientId: "1060725074195-kmeum4crr01uirfl2op9kd5acmi9jutn.apps.googleusercontent.com",
+                    clientSecret: "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",
+                }),
+                ThirdPartyEmailPassword.Github({
+                    clientSecret: "e97051221f4b6426e8fe8d51486396703012f5bd",
+                    clientId: "467101b197249757c71f",
+                }),
+                ThirdPartyEmailPassword.Apple({
+                    clientId: "4398792-io.supertokens.example.service",
+                    clientSecret: {
+                        keyId: "7M48Y4RYDL",
+                        privateKey:
+                            "-----BEGIN PRIVATE KEY-----\nMIGTAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBHkwdwIBAQQgu8gXs+XYkqXD6Ala9Sf/iJXzhbwcoG5dMh1OonpdJUmgCgYIKoZIzj0DAQehRANCAASfrvlFbFCYqn3I2zeknYXLwtH30JuOKestDbSfZYxZNMqhF/OzdZFTV0zc5u5s3eN+oCWbnvl0hM+9IW0UlkdA\n-----END PRIVATE KEY-----",
+                        teamId: "YWQCXGJRJL",
+                    },
+                }),
+            ],
+        }),
+        Session.init({
+            jwt: {
+                enable: true,
+                issuer: process.env.JWT_ISSUER_URL,
+            },
+            override: {
+                functions: (oI) => {
+                    return {
+                        ...oI,
+                        createNewSession: async function (input) {
+                            let userInfo = await ThirdPartyEmailPassword.getUserById(input.userId);
+
+                            input.accessTokenPayload = {
+                                ...input.accessTokenPayload,
+                                "https://hasura.io/jwt/claims": {
+                                    "x-hasura-user-id": userInfo.email,
+                                    "x-hasura-default-role": "user",
+                                    "x-hasura-allowed-roles": ["user"],
+                                },
+                            };
+
+                            return oI.createNewSession(input);
+                        },
+                    };
+                },
+            },
+        }),
+    ],
+});
+
+const app = express();
+
+app.use(
+    cors({
+        origin: websiteDomain, // TODO: Change to your app's website domain
+        allowedHeaders: ["content-type", ...supertokens.getAllCORSHeaders()],
+        methods: ["GET", "PUT", "POST", "DELETE"],
+        credentials: true,
+    })
+);
+
+app.use(
+    helmet({
+        contentSecurityPolicy: false,
+    })
+);
+app.use(middleware());
+
+// Application APIs are exposed in Hasura
+
+app.use(errorHandler());
+
+app.use((err, req, res, next) => {
+    res.status(500).send("Internal error: " + err.message);
+});
+
+app.listen(apiPort, () => console.log(`API Server listening on port ${apiPort}`));